% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zcta_crosswalk}
\alias{get_zcta_crosswalk}
\title{Returns a ZCTA Crosswalk as a tibble}
\usage{
get_zcta_crosswalk(url = "https://rb.gy/h0l5cs", process_file = TRUE)
}
\arguments{
\item{url}{The URL of the file to parse. Defaults to the 2020 ZCTA to County Relationship File}

\item{process_file}{If TRUE, remove NA rows, unused columns, etc.}
}
\description{
By default, returns the Census Bureau's 2010 ZCTA Country Relationship file
as a tibble.
}
\seealso{
Zip Code Tabulation Area 5-Digit (ZCTA5) Relationship Files: https://rb.gy/h0l5cs
}
