% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zcta_metadata}
\alias{get_zcta_metadata}
\title{Return metadata on a ZCTA}
\usage{
get_zcta_metadata(zctas)
}
\arguments{
\item{zctas}{A vector of ZCTAs (character or numeric)}
}
\value{
A tibble, where each row is a (zcta, county, state) combination.
}
\description{
Given a vector of ZIP Code Tabulation Areas (ZCTAs), return what
state and county they are in. NOTE: A single ZCTA can span multiple
states and counties.
}
\examples{
get_zcta_metadata("90210")

# Some ZCTAs span multiple counties
get_zcta_metadata(39573)
}
