% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zctas_in_county}
\alias{get_zctas_in_county}
\title{Return the ZCTAs in a vector of counties}
\usage{
get_zctas_in_county(counties)
}
\arguments{
\item{counties}{A vector of Counties as FIPS codes. Must be 5-digits as characters - see examples.}
}
\description{
Given a vector of counties, return the Zip Code Tabulation Areas (ZCTAs)
in those counties
}
\examples{
# 06075 is San Francisco County, California
get_zctas_in_county("06075")

# "36059" is Nassau County, New York
get_zctas_in_county(c("06075", "36059"))
}
