% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zctas_in_state}
\alias{get_zctas_in_state}
\title{Return the ZCTAs in a vector of states}
\usage{
get_zctas_in_state(state_fips)
}
\arguments{
\item{state_fips}{A vector of Counties as FIPS codes. Must be 2-digits as characters - see examples.}
}
\description{
Given a vector of states, return the Zip Code Tabulation Areas (ZCTAs)
in those states
}
\examples{
# 06 is California
ca_zctas = get_zctas_in_state("06")
length(ca_zctas)
head(ca_zctas)

# "36" is New York
ny_ca_zctas = get_zctas_in_state(c("36", "06"))
length(ny_ca_zctas)
head(ny_ca_zctas)
}
