% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_names}
\alias{state_names}
\title{Metadata for Each "State" in zcta_crosswalk}
\usage{
data(state_names)
}
\description{
The complete dataset in ?zcta_crosswalk contains information on 56 state and
state-equivalents. This dataframe contains the full name of each "state", plus
its USPS abbreviation and FIPS code.
}
