% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_crosswalk}
\alias{zcta_crosswalk}
\title{2020 Crosswalk of ZIP Code Tabulation Areas (ZCTAs)}
\usage{
data(zcta_crosswalk)
}
\description{
Retrieved via the function get_zcta_crosswalk. Columns are ZCTA, county_fips,
county_name and state_fips.
}
