% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_cw}
\alias{ct_cw}
\title{Crosswalk between Connecticut County Equivalents (Planning Regions) and ZCTAs}
\usage{
data(ct_cw)
}
\description{
The complete dataset in ?ct_cw contains information Connecticut's new planning
regions and their corresponding ZCTAs, names, and GEOIDs. This data was sourced
from the U.S. Census Bureau's TIGER/Line shapefiles pulled with the tigris package.
Intersections were calculated using the sf package. Any ZCTA that crosses into
a planning region's border is included in that planning region's associated ZCTAs.
}
\details{
The addition of this data structure attempts to resolve issues in the Census Bureau's
ZCTA relationship files that came about due to the State of Connecticut's change
to county-equivalents (planning regions). This change was implemented in January 2022 and
is further documented by the Census Bureau's
\href{https://www.census.gov/programs-surveys/geography/technical-documentation/county-changes.2020.html#list-tab-957819518}{county changes page}
and the corresponding
\href{https://www.federalregister.gov/documents/2022/06/06/2022-12063/change-to-county-equivalents-in-the-state-of-connecticut}{Federal Register notice.}
}
