% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_cw}
\alias{ct_cw}
\title{Crosswalk between Connecticut County Equivalents (Planning Regions) and ZCTAs}
\usage{
data(ct_cw)
}
\description{
The complete dataset in ?ct_cw contains information Connecticut's new planning
regions and their corresponding ZCTAs, names, and GEOIDs. This data was sourced
from the U.S. Census Bureau's TIGER/Line shapefiles pulled with the tigris package.
Intersections were calculated using the sf package. Any ZCTA that crosses into
a planning region's border is included in that planning region's associated ZCTAs.
}
