% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zcta_crosswalk}
\alias{get_zcta_crosswalk}
\title{Returns a ZCTA Crosswalk as a tibble}
\usage{
get_zcta_crosswalk()
}
\description{
Returns the Census Bureau's 2020 ZCTA Country Relationship file
as a tibble.
}
\seealso{
All 2020 Zip Code Tabulation Area 5-Digit (ZCTA5) Relationship Files: https://rb.gy/h0l5cs
}
