% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zcta_crosswalk}
\alias{get_zcta_crosswalk}
\title{Returns a ZCTA crosswalk as a tibble}
\usage{
get_zcta_crosswalk()
}
\value{
A tibble, where each row is a (zcta, county, state) combination.
}
\description{
Returns the Census Bureau's 2020 ZCTA Country Relationship file
as a tibble. This function is included so that users can see how the crosswalk
was generated. It is not intended for use by end users.
}
\seealso{
All 2020 ZIP Code Tabulation Area 5-Digit (ZCTA5) Relationship Files: https://rb.gy/h0l5cs
}
