% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcta.R
\name{get_zcta_metadata}
\alias{get_zcta_metadata}
\title{Return metadata on a ZCTA}
\usage{
get_zcta_metadata(zctas)
}
\arguments{
\item{zctas}{A vector of ZCTAs}
}
\description{
Given a vector of ZIP Code Tabulation Areas (ZCTAs), return what
state and county they are in. NOTE: A single ZCTA can span multiple
states and counties.
}
\examples{
get_zcta_metadata("90210")

get_zcta_metadata("39573")
}
