% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta_crosswalk}
\alias{zcta_crosswalk}
\title{2020 Crosswalk of ZIP Code Tabulation Areas (ZCTAs)}
\usage{
data(zcta_crosswalk)
}
\description{
The primary data was obtained via the function get_zcta_crosswalk. There are 3 types of columns: ZCTA,
state and county. Where data in practice sometimes appears as both character and numeric, columns for both are 
provided.
}
